+++ lib/src/listener/signaling_listener.dart
@@ -10,6 +10,8 @@ class OnSignalingListener {
   final Function(SignalingInfo info)? onInviteeRejectedByOtherDevice;
   final Function(SignalingInfo info)? onInviteeAcceptedByOtherDevice;
   final Function(SignalingInfo info)? onHangup;
+  final Function(RoomCallingInfo info)? onRoomParticipantConnected;
+  final Function(RoomCallingInfo info)? onRoomParticipantDisconnected;
 
   OnSignalingListener({
     this.onInvitationCancelled,
@@ -20,6 +22,8 @@ class OnSignalingListener {
     this.onInviteeAcceptedByOtherDevice,
     this.onInviteeRejectedByOtherDevice,
     this.onHangup,
+    this.onRoomParticipantConnected,
+    this.onRoomParticipantDisconnected,
   });
 
   /// 被邀请者收到：邀请者取消音视频通话
@@ -60,5 +64,13 @@ class OnSignalingListener {
   /// 被挂断
   void hangup(SignalingInfo info) {
     onHangup?.call(info);
+  }
+
+  void roomParticipantConnected(RoomCallingInfo info) {
+    onRoomParticipantConnected?.call(info);
+  }
+
+  void roomParticipantDisconnected(RoomCallingInfo info) {
+    onRoomParticipantDisconnected?.call(info);
   }
 }